pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "amrcast"
version = "0.1.0"
description = "ML-powered antibiotic resistance prediction with quantitative MIC forecasting"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "kglazier" }]
keywords = ["antimicrobial-resistance", "MIC", "bioinformatics", "machine-learning", "genomics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9,<1",
    "rich>=13.0",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "biopython>=1.80",
    "pyrodigal>=3.0",
    "pyhmmer>=0.10",
    "numpy>=1.24,<2",
    "pandas>=2.0,<3",
    "xgboost>=2.0,<3",
    "scikit-learn>=1.3",
    "shap>=0.42",
    "requests>=2.28",
    "pyyaml>=6.0",
    "h5py>=3.9",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.0,<3",
    "fair-esm>=2.0",
]
api = [
    "fastapi>=0.100",
    "uvicorn>=0.23",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.5",
    "pre-commit>=3.0",
]

[project.scripts]
amrcast = "amrcast.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/amrcast"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
